etl2/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
geographiclib==2.0
geopy==2.4.0
greenlet==3.0.0
numpy==1.26.1
pandas==2.1.1
prompt-toolkit==3.0.36
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
questionary==2.0.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
termcolor==2.3.0
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.8
